Refactor and add new functions to limit db open time
Only functions that directly interact with (i.e. query) the database will open the connection and then close immediately after. This will allow issues with the db staying open, especially when an error is encountered. 100 open and closes of the db takes ~5 seconds. If the user wants the connection to stay open, there is a stay_open attribute that is settable on init or through available methods.